keyball44/daniel2887: Add anisotropy support
This feature applies a per-axis multiplier, prior to further
post-processing like acceleration. The intent is to weigh X and Y
movements differently, to compensate for biomechanical nature of the
thumb and the trackball's position on Keyball keyboards.

The feature itself is simple. Most of this change is refactoring related
to keyball configuration.